Add Colley to aggregate MOV
I've been not in agreement with the rest of the computer rankings for much of this year using RW.  Instead use the mean field approach of the Colley Matrix to knock down noise?

Also fixed all the scoring accounting for everything over 1% in D3.  Still need to look at D1.